Package restore working

+++ src/build.sh
@@ -1,17 +1,31 @@
-# download nuget command line
-curl https://dist.nuget.org/win-x86-commandline/latest/nuget.exe > nuget.exe
+#!/bin/bash
+set -e
 
-# run it (don't need to make it executable this way)
-mono nuget.exe restore
+# download nuget command line if not present already
+# if [ ! -e nuget.exe ]; then
+#     echo "Downloading latest nuget.exe"
+#     curl https://github.com/mono/nuget-binary/blob/master/nuget.exe > nuget.exe
+#     #curl https://dist.nuget.org/win-x86-commandline/latest/nuget.exe > nuget.exe
+# fi
 
-# build the solution
+# no - use apt-get package now!
+# https://stackoverflow.com/questions/38118548/how-to-install-nuget-from-command-line-on-linux
+
+# run it
+if [ ! -d packages ]; then
+    echo "Package restore..."
+    # mono --runtime=v4.0 nuget.exe restore PiCandy.sln
+    nuget restore PiCandy.sln
+fi
+
+if [ ! -e "/lib/libws2011.so" ]; then
+    echo "Building libws2011.so (dependency)"
+    pushd rpi_ws281x/mono
+    ./build.sh
+    sudo cp ../libws2811.so /lib
+    popd
+fi
+
+echo "Building solution"
+# build the solution (nb: not yet using FAKE here)
 xbuild
-
-# if no libws2011.so need to build that too
-# Build and copy the WS281x driver to lib
-pushd ../
-git clone https://github.com/piers7/rpi_ws281x
-cd rpi_ws281x/mono
-./build.sh
-sudo cp ../libws2811.so /lib
-popd
